chore(refactor): cleanup imports and exports
