+++ packages/lib/src/CheckResult.ts
